markups - small renames and restyle, split files
